% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Jacobians_Functions.R
\name{assemble_jacobian}
\alias{assemble_jacobian}
\title{Constructs a community matrix from the data.frame created by the interaction_strengths function}
\usage{
assemble_jacobian(df, species_list)
}
\arguments{
\item{df}{the dataframe created by the interaction strengths function}

\item{species_list}{list of species names in the community, can be taken from the abundance table}
}
\value{
The community matrix (so far as a data.frame)
}
\description{
To do this, the list of species names from the abundace table is used to name the rows and columns
of the community matrix. The corresponding interaction strengths are picked from the dataframe, based on
the species name.
}
