% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_s.R
\name{find_s}
\alias{find_s}
\title{Find the relative amount of self-regulation needed for stability s*}
\usage{
find_s(Jacobian, step_size = 0.01, start_s = 1, max_s = 1000)
}
\arguments{
\item{Jacobian}{A Jacobian matrix}

\item{step_size}{Amount by which \emph{s} is increases in each step}

\item{start_s}{Starting value of \emph{s}}

\item{max_s}{Maximum possible \emph{s} value. If the matrix is still not stable, NA is returned}
}
\description{
For unstable matrices, diagonal values are incrementally increased by multiplying it
with a factor \emph{s}, until its dominant eigenvalue becomes negative (=the matrix becomes stable).
}
\details{
This will not work, if the diagonal contains zeros!
Replace missing values before calculating s*!

For stable matrices, diagonal values are incrementally decrease, until the
dominant eigenvalue becomes positive (= the matrix is unstable)
}
