% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_s.R
\name{find_s_F}
\alias{find_s_F}
\title{Find the relative amount of self-regulation needed to fullfill Levins I
stability criterium}
\usage{
find_s_F(Jacobian, step_size = 0.01, max_s = 1000)
}
\arguments{
\item{Jacobian}{A Jacobian matrix}

\item{step_size}{Amount by which s is increases in each step}

\item{max_s}{Maximum possible s value. If the matrix is still not stable, NA is returned}
}
\description{
Diagonal values are incrementally increases by a factor s until all total feedback
values (calculated via the coefficients of the characteristic polynomial).
This will not work, if the diagonal contains zeros.
Replace missing values before calculating *s\**!
}
