% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_feedback.R
\name{get_Fk}
\alias{get_Fk}
\title{Calculates total feedback at a given level \emph{k}}
\usage{
get_Fk(k, A)
}
\arguments{
\item{k}{level for which total feedback should be calculated}

\item{A}{a Jacobian matrix}
}
\value{
F_k total feedback at level k
}
\description{
F_k is calculated via the coefficients of the characteristic polynomial
}
