% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_names}
\alias{get_names}
\title{Helper function to get data set names from the path of csv file}
\usage{
get_names(path_list, string)
}
\arguments{
\item{path_list}{List of paths to csv files (one for each dataset)}

\item{string}{The parts of the file name that doesn't belong to the name, e.g. "abundance_cleaned"}
}
\value{
a list of data set file names
}
\description{
returns filenames with .csv ending
}
