% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_names_sans}
\alias{get_names_sans}
\title{Helper functions to get data set names}
\usage{
get_names_sans(path_list, string)
}
\arguments{
\item{path_list}{List of paths to csv files}

\item{string}{The parts of the file name that doesn't belong to the name, e.g. "Abundance_cleaned"}
}
\value{
a list of data set names
}
\description{
returns filenames without .csv ending
}
