% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops_functions_new.R
\name{loop_weight}
\alias{loop_weight}
\title{Calculate the loop strength and weight of a given feedback loop}
\usage{
loop_weight(loop, A)
}
\arguments{
\item{loop}{a vector containing the species that are in the loop, e.g. c(1,2,3)
for the 3-link loop 1 -> 2 -> 3}

\item{A}{A Jacobian matrix}
}
\value{
A vector containing the loop strength and the loop weight
}
\description{
Used internally by loops().
}
\details{
A loop is defined by a list of species e.g. c(1,2,3) is the 3-link loop 1 -> 2 -> 3
these species numbers are used as indices in order to pick the corresponding
interaction strengths from the Jacobian \eqn{\mathbf{A}}
}
