% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops_functions_new.R
\name{loops}
\alias{loops}
\title{Calculate the strengths and weights of all loops of length \emph{n} within the matrix \emph{A}}
\usage{
loops(n, A)
}
\arguments{
\item{n}{loop length}

\item{A}{a Jacobian matrix in which all loops of length should be identified}
}
\value{
A dataframe containing the loops, their strengths and weights
}
\description{
To do this, I simply assume that all possible loops of length n (all possible
combinations of n-species) exist. Then, the function goes through all possible
loops and calculates their strengths using loop_weights().
}
\details{
If the loop weight is 0, at least one of the links is missing and the loop
is removed from the list.
}
