% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops_functions_new.R
\name{loops_parallel}
\alias{loops_parallel}
\title{Calculates the strengths and weights of all loops of length \emph{n} within the matrix \emph{A}}
\usage{
loops_parallel(n, A, n_cores)
}
\arguments{
\item{n}{loop length}

\item{A}{a Jacobian matrix in which all loops of length should be identified}

\item{n_cores}{number of cores used for computation}
}
\value{
A dataframe containing the loops, their strengths and weights
}
\description{
Same function as loops() but calculations can be run on multiple cores
(setup of parallel backend only works on Linux!!!)
}
