% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomization_functions.R
\name{randomize_all}
\alias{randomize_all}
\title{Full randomisation of an interaction table}
\usage{
randomize_all(df, ij_col, ji_col)
}
\arguments{
\item{df}{interaction table (created by interaction_strenghts())}

\item{ij_col}{column of \eqn{a_{ij}} values to randomise (choose scaled or unscaled)}

\item{ji_col}{column of \eqn{a_{ji}} values to randomise (scaled or unscaled)}
}
\value{
interaction table with two additional columns \emph{$a_ij_rand}
and \emph{$a_ji_rand} that contain the interaction strengths in a randomised order
}
\description{
During the full randomisation procedure, ALL interspecific interaction strengths are randomly
reshuffled within the network. Intraspecific interactions, the diagonal matrix elements
are not affects and network topology is preserved (zeros remain in place).
}
\details{
The function returns a new interaction table that contains two new columns $a_ij_rand and
$a_ji_rand that contains the same values as the original columns but in a different order.

To get a fully randomised Jacobian matrix, use assemble_jacobian() and specify
the new columns.
}
