% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_contact_matrix}
\alias{read_contact_matrix}
\title{Reads in the species-contact matrix and returns a nicely formatted data frame
that is more easy to work with in R.}
\usage{
read_contact_matrix(path_competition)
}
\arguments{
\item{path_competition}{path to species contact matrix .csv file}
}
\value{
a data frame containing the species contact matrix
}
\description{
Reads in the species-contact matrix and returns a nicely formatted data frame
that is more easy to work with in R.
}
