% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Jacobians_Functions.R
\name{read_data}
\alias{read_data}
\title{Read in the raw data files}
\usage{
read_data(path_competition, path_abundance)
}
\arguments{
\item{path_competition}{path to species contact matrix .csv file}

\item{path_abundance}{path to abundance .csv file}
}
\value{
A list containing two data frames: the
species contact matrix and a cleaned abundance table
}
\description{
Competition and abundance data must be read in together, to make sure that species match.
Some species may appear in the abundance table but not in the species contact matrix,
because they do not interact with anyone yet). These species are dropped from the list.
Species in the abundance table are also reordered to match the order of species
in the species contact matrix.
}
