% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{replace_zeros}
\alias{replace_zeros}
\title{Replace missing diagonal values}
\usage{
replace_zeros(Jac, f)
}
\arguments{
\item{Jac}{A Jacobian matrix}

\item{f}{factor multiplied with mean interaction strength to choose replacement value}
}
\value{
A new version of the Jacobian in which missing diagonal values have been replaced
}
\description{
The scaling procedure as well as the calculation of s* require that all
diagonal values $a_{ii} != 0$. Therefore, missing values need to be replaced by
some small value.
}
\details{
This replacement value is calculated as the mean strengths of all off-diagonal
matrix elements, mutltiplied with a factor f.
}
