% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale interaction strengths within a Jacobian matrix}
\usage{
scale_matrix(A, aii = 0)
}
\arguments{
\item{A}{A Jacobian matrix, with non-zero diagonal elements}

\item{aii}{determines the value of diagonal elements}
}
\value{
The same Jacobian matrix with scaled interaction strengths and -1 on the diagonal
}
\description{
Off-diagonal values are scaled by dividing them by the absolute value of
their corresponding diagonal element. The resulting scaled interaction strengths
are thus defined as: \eqn{\bar{a_{ij}} = \frac{a_{ij}}{|a_{ii}|}} (Neutel and Thorne
2016).
}
\details{
This does not work if there is a 0 on the diagonal ! Replace missing elements
with replace_zeros() before scaling !.
}
