% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Jacobians_Functions.R
\name{interaction_strengths}
\alias{interaction_strengths}
\title{Calculates interaction_strengths from species_contact matrix}
\usage{
interaction_strengths(competition, abundance, cost_list)
}
\arguments{
\item{competition}{species-contact matrix}

\item{abundance}{abundance table}

\item{cost_list}{list containing win_cost, loss_cost, draw_cost}
}
\value{
table containing biomass loss rates and interaction strenghts
}
\description{
In a first step, the species-contact matrix is turned into a tabular
format with one row per pairwise interaction in order to make computations easier. Then,
biomass loss rates are calculated with a given list of cost-values. Biomass loss rates are then
converted to interaction strengths by diving them by the respective species' abundance.
}
