% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Jacobians_Functions.R
\name{interaction_strengths}
\alias{interaction_strengths}
\title{Calculate interaction strengths}
\usage{
interaction_strengths(competition, abundance, cost_list)
}
\arguments{
\item{competition}{species-contact matrix}

\item{abundance}{abundance table}

\item{cost_list}{list containing win_cost, loss_cost, draw_cost}
}
\value{
A dataframe containing biomass loss rates and interaction strengths
}
\description{
Turns the species-contact matrix into a tabular format with one row per
pairwise interaction. Biomass loss rates are calculated with a given list
of cost-values. Interaction strengths are calculated by dividing biomass loss
rate \eqn{B_{ij}} by the abundance of species \eqn{B_j}.
}
