% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomization_functions.R
\name{randomize_all}
\alias{randomize_all}
\title{Full randomisation of an interaction table}
\usage{
randomize_all(df)
}
\arguments{
\item{df}{interaction table (created by interaction_strenghts())}
}
\value{
the same interaction table but with two additional columns $ F_ij_B_rand
and $F_ji_B_rand that contain the same interaction strengths in a randomised order
}
\description{
During the full randomisation procedure, ALL interspecific interaction strengths are randomly
reshuffled within the network. Intraspecific interactions/ Diagonal values are not affected.
interaction columns so that an F_ij value can become an F_ji value and
vice-vers. The function returns a new interaction table
that contains two columns with randomised interaction strengths.
Use this table as input
to assemble_jacobian_randomised() to get a fully randomised Jacobian matrix.
Note: values are only randomised across existing links! (the interaction table
does not contain non-existent links). Thus, if the Jacobian is reconstructed from
the randomised interaction table, the network will have the same topology as
the empirical one.
}
