% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomization_functions.R
\name{randomize_pw}
\alias{randomize_pw}
\title{Pairiwise randomisation of an interaction table}
\usage{
randomize_pw(df)
}
\arguments{
\item{df}{interactiontable (created by interaction_strenghts())}
}
\value{
the same interaction table with additional columns $F_ij_B_pw and $F_ji_B_pw,
containing randomised interaction strengths
}
\description{
During pairwise randomisation, pairs of interaction strengths are kept intact but their
location is reshuffled across the network. In the interaction table this means the following:
F_ij and F_ji values that appear in the same row in the original table, will also be in the same
row in the pairwise randomised table(all though it is possible that they switch columns).
The function returns a new interaction table that contains two columns with randomised interaction strengths,
Use this table as input
to assemble_jacobian_randomised() to get a fully randomised Jacobian matrix.
Note: values are only randomised across existing links! (the interaction table
does not contain non-existent links). Thus, if the Jacobian is reconstructed from
the randomised interaction table, the network will have the same topology as
the empirical one!
}
