% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Jacobians_Functions.R
\name{read_data}
\alias{read_data}
\title{Reads in the raw data files and returns data frames that are more easy to work with in R.
Competition and abundance data must be read in together, to make sure the species match.}
\usage{
read_data(path_competition, path_abundance)
}
\arguments{
\item{path_competition}{path to species contact matrix .csv file}

\item{path_abundance}{path to abundance .csv file}
}
\value{
A list containing a competition data frame, that contains the
species contact matrix and acleaned abundance table (missing species are omitted +
same order as competition table)
}
\description{
Reads in the raw data files and returns data frames that are more easy to work with in R.
Competition and abundance data must be read in together, to make sure the species match.
}
